exemplo_autenticacao/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2-binary
requests
django-filter